# docs/swagger.yaml
openapi: 3.0.4
info:
  title: QRead.backend - OpenAPI 3.0
  description: |-
    Backend for our QRead FYP project
  version: "1.0"
...
servers:
  - url: http://127.0.0.1
...
paths:
...
  /admin/users:
    get:
      summary: Retrieve users by search
      tags:
        - Admin
      security:
        - qread.backend_auth: []
      parameters:
        - in: query
          name: id
          schema:
            type: integer
          required: false
          description: User's id
        - in: query
          name: name
          schema:
            type: string
          required: false
          description: User's name
        - in: query
          name: email
          schema:
            type: string
          required: false
          description: User's email
        - in: query
          name: type
          schema:
            type: string
            format: enum
            enum: ["ADMIN", "LIBRARIAN", "BORROWER"]
          required: false
          description: User's account type
        - in: query
          name: state
          schema:
            type: string
            format: enum
            enum: ["ACTIVE", "SUSPENDED"]
          required: false
          description: User's account state
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Users retrieved
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 12
                        name:
                          type: string
                          example: name
                        email:
                          type: string
                          example: test@email.com
                        account_type:
                          type: string
                          enum: [ADMIN, LIBRARIAN, BORROWER]
                          example: LIBRARIAN
                        account_state:
                          type: string
                          enum: [ACTIVE, SUSPENDED]
                          example: ACTIVE
                    example:
                      - user:
                          id: 12
                          name: admin_name
                          email: admin@email.com
                          account_type: ADMIN
                          account_state: ACTIVE
                        transaction:
                          - id: 21
                            user_id: 12
                            book_id: 39
                            date: 2025-08-17T19:58:00Z
                            due: 2025-08-20T19:58:00Z
                            returned: False
                          - id: 22
                            user_id: 12
                            book_id: 40
                            date: 2025-08-17T19:58:00Z
                            due: 2025-08-20T19:58:00Z
                            returned: False
                        fine:
                          - id: 33
                            user_id: 12
                            transaction_id: 21
                            reason: Late return
                            amount: 50
                            date: 2025-08-17T19:58:00Z
                            paid: False
        "401":
          description: Authentication error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Unauthorized"
  /admin/user:
    delete:
      summary: Delete a user
      tags:
        - Admin
      security:
        - qread.backend_auth: []
      parameters:
        - in: query
          name: id
          schema:
            type: string
          required: true
          description: The user's id to be deleted
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User deleted successfully"
        "400":
          description: Error operations
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Missing id field
              examples:
                Missing id:
                  value:
                    error: Missing id field
                Invalid id:
                  value:
                    error: Invalid id 49
        "401":
          description: Authentication error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Unauthorized"
        "404":
          description: User id not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: User id 48 not found
  /admin/suspend-user:
    post:
      summary: Suspend a user
      tags:
        - Admin
      security:
        - qread.backend_auth: []
      requestBody:
        required: True
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  example: 49
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User suspended successfully"
        "400":
          description: Error operations
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Missing id field
              examples:
                Missing id:
                  value:
                    error: Missing id field
                Invalid id:
                  value:
                    error: Invalid id 49
        "401":
          description: Authentication error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Unauthorized"
        "404":
          description: User id not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: User id 48 not found
  /admin/reinstate-user:
    post:
      summary: Reinstate a user
      tags:
        - Admin
      security:
        - qread.backend_auth: []
      requestBody:
        required: True
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  example: 49
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User reinstated successfully"
        "400":
          description: Error operations
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Missing id field
              examples:
                Missing id:
                  value:
                    error: Missing id field
                Invalid id:
                  value:
                    error: Invalid id 49
        "401":
          description: Authentication error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Unauthorized"
        "404":
          description: User id not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: User id 48 not found
  /admin/register-librarian:
    post:
      summary: Register a new librarian
      description: An email will be sent to the librarian which should be followed by an API call to /new-librarian
      tags:
        - Admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: test@email.com
      security:
        - qread.backend_auth: []
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Invitation email sent"
        "400":
          description: Error operations
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Missing email field
              examples:
                Missing email:
                  value:
                    error: Missing email field
                Invalid email:
                  value:
                    error: Invalid email test.com
                Email already exist:
                  value:
                    error: Email already exist
        "401":
          description: Authentication error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Unauthorized
  /admin/app-setting:
    put:
      summary: Update a setting
      tags:
        - Admin
      security:
        - qread.backend_auth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                key:
                  type: string
                  example: reminder_x_days_before_due
                value:
                  type: integer
                  example: 5
      responses:
        "200":
          description: App setting updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Setting updated successfully"
        "400":
          description: Key not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Key random_key not found"
              examples:
                Missing key:
                  value:
                    error: Missing key field
                Missing value:
                  value:
                    error: Missing value field
        "401":
          description: Authentication error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Unauthorized"
        "404":
          description: Key not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Key random_key not found"
components:
  securitySchemes:
    qread.backend_auth:
      type: apiKey
      in: cookie
      name: session